requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
evaluate==0.4.3
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2024.5.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==15.0.2
pyarrow-hotfix==0.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.20.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
